polkavm-linker: move `section_to_function_name` creation to separate function
